fix(resumable): PendingPack.remaining_count u8 -> u32 to match decoded.sym_count

+++ src/decompress/inflate/resumable.rs
@@ -73,7 +73,7 @@ pub struct PendingMatch {
 #[derive(Copy, Clone, Debug)]
 pub struct PendingPack {
     pub remaining_symbol: u32,
-    pub remaining_count: u8,
+    pub remaining_count: u32,
 }
 
 /// Block-decode state machine position. Persists across `read_stream` calls
